WIP: Path tracking vehicle elimination bug Consolidated path tracking code in LocalizationUtils.h/.cpp Moved utility functions to LocalizationUtils.h/.cpp Removed TrafficDistributor.h/.cpp
